requirements.txt
flask
beautifulsoup4
requests
selenium
mongoengine
python-dotenv
gunicorn
lxml